+++ Sources/TuistKit/Models/Project.swift
@@ -2,7 +2,7 @@ import Basic
 import Foundation
 import TuistCore
 
-class Project: Equatable {
+class Project: Equatable, CustomStringConvertible {
     // MARK: - Attributes
 
     /// Path to the folder that contains the project manifest.
@@ -72,6 +72,47 @@ class Project: Equatable {
         targets = try targetsJSONs.map({ try Target(dictionary: $0, projectPath: path, fileHandler: fileHandler) })
         let settingsJSON: JSON? = try? json.get("settings")
         settings = try settingsJSON.map({ try Settings(dictionary: $0, projectPath: path, fileHandler: fileHandler) })
+    }
+    
+    /// It returns the project targets sorted based on the target type and the dependencies between them.
+    /// The most dependent and non-tests targets are sorted first in the list.
+    ///
+    /// - Parameter graph: Dependencies graph.
+    /// - Returns: Sorted targets.
+    func sortedTargetsForProjectScheme(graph: Graphing) -> [Target] {
+        return targets.sorted { (first, second) -> Bool in
+            // First criteria: Test bundles at the end
+            if first.product.testsBundle && !second.product.testsBundle {
+                return false
+            }
+            if !first.product.testsBundle && second.product.testsBundle {
+                return true
+            }
+            
+            // Second criteria: Most dependent targets first.
+            let secondDependencies = graph.targetDependencies(path: self.path, name: second.name)
+                .filter({ $0.path == self.path })
+                .map({ $0.target.name })
+            let firstDependencies = graph.targetDependencies(path: self.path, name: first.name)
+                .filter({ $0.path == self.path })
+                .map({ $0.target.name })
+            
+            if secondDependencies.contains(first.name) {
+                return true
+            } else if firstDependencies.contains(second.name) {
+                return false
+                
+                // Third criteria: Name
+            } else {
+                return first.name < second.name
+            }
+        }
+    }
+    
+    // MARK: - CustomStringConvertible
+    
+    var description: String {
+        return self.name
     }
 
     // MARK: - Equatable