+++ Tests/TuistKitTests/Generator/Mocks/MockSchemesGenerator.swift
@@ -3,8 +3,13 @@ import Foundation
 
 final class MockSchemesGenerator: SchemesGenerating {
     var generateTargetSchemesArgs: [(project: Project, generatedProject: GeneratedProject)] = []
+    var generateProjectSchemeArgs: [(project: Project, generatedProject: GeneratedProject, graph: Graphing)] = []
 
     func generateTargetSchemes(project: Project, generatedProject: GeneratedProject) throws {
         generateTargetSchemesArgs.append((project: project, generatedProject: generatedProject))
+    }
+    
+    func generateProjectScheme(project: Project, generatedProject: GeneratedProject, graph: Graphing) throws {
+        generateProjectSchemeArgs.append((project: project, generatedProject: generatedProject, graph: graph))
     }
 }