+++ Tests/TuistKitTests/Generator/SchemesGeneratorTests.swift
@@ -14,28 +14,91 @@ final class SchemeGeneratorTests: XCTestCase {
         super.setUp()
         subject = SchemesGenerator()
     }
+    
+    func test_projectBuildAction() {
+        let app = Target.test(name: "App", product: .app)
+        let appTests = Target.test(name: "AppTests", product: .unitTests)
+        let appUITests = Target.test(name: "AppUITests", product: .uiTests)
+        let targets = [app, appTests, appUITests]
 
-    func test_testAction_when_notTestsTarget() {
+        let project = Project.test(targets: targets)
+        let graphCache = GraphLoaderCache()
+        let graph = Graph.test(cache: graphCache)
+        
+        let got = subject.projectBuildAction(project: project,
+                                            generatedProject: generatedProject(targets: targets),
+                                            graph: graph)
+        
+        XCTAssertTrue(got.parallelizeBuild)
+        XCTAssertTrue(got.buildImplicitDependencies)
+        XCTAssertEqual(got.buildActionEntries.count, 3)
+        
+        let appEntry = got.buildActionEntries[0]
+        let testsEntry = got.buildActionEntries[1]
+        let uiTestsEntry = got.buildActionEntries[2]
+        
+        XCTAssertEqual(appEntry.buildFor, [.analyzing, .archiving, .profiling, .running, .testing])
+        XCTAssertEqual(appEntry.buildableReference.referencedContainer, "container:project.xcodeproj")
+        XCTAssertEqual(appEntry.buildableReference.buildableName, app.productName)
+        XCTAssertEqual(appEntry.buildableReference.blueprintName, app.name)
+        XCTAssertEqual(appEntry.buildableReference.buildableIdentifier, "primary")
+
+        XCTAssertEqual(testsEntry.buildFor, [.testing])
+        XCTAssertEqual(testsEntry.buildableReference.referencedContainer, "container:project.xcodeproj")
+        XCTAssertEqual(testsEntry.buildableReference.buildableName, appTests.productName)
+        XCTAssertEqual(testsEntry.buildableReference.blueprintName, appTests.name)
+        XCTAssertEqual(testsEntry.buildableReference.buildableIdentifier, "primary")
+        
+        XCTAssertEqual(uiTestsEntry.buildFor, [.testing])
+        XCTAssertEqual(uiTestsEntry.buildableReference.referencedContainer, "container:project.xcodeproj")
+        XCTAssertEqual(uiTestsEntry.buildableReference.buildableName, appUITests.productName)
+        XCTAssertEqual(uiTestsEntry.buildableReference.blueprintName, appUITests.name)
+        XCTAssertEqual(uiTestsEntry.buildableReference.buildableIdentifier, "primary")
+    }
+    
+    func test_projectTestAction() {
+        let app = Target.test(name: "App", product: .app)
+        let appTests = Target.test(name: "AppTests", product: .unitTests)
+        let targets = [app, appTests]
+        let project = Project.test(targets: targets)
+        
+        let got = subject.projectTestAction(project: project,
+                                            generatedProject: generatedProject(targets: targets))
+        
+        XCTAssertEqual(got.buildConfiguration, "Debug")
+        XCTAssertNil(got.macroExpansion)
+        XCTAssertEqual(got.testables.count, 1)
+        
+        let testable = got.testables.first
+        XCTAssertEqual(testable?.skipped, false)
+        
+        XCTAssertEqual(testable?.buildableReference.referencedContainer, "container:project.xcodeproj")
+        XCTAssertEqual(testable?.buildableReference.buildableName, appTests.productName)
+        XCTAssertEqual(testable?.buildableReference.blueprintName, appTests.name)
+        XCTAssertEqual(testable?.buildableReference.buildableIdentifier, "primary")
+    }
+
+    func test_targetTestAction_when_notTestsTarget() {
         let target = Target.test(name: "AppTests", product: .app)
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
-
-        let got = subject.testAction(target: target,
-                                     pbxTarget: pbxTarget,
-                                     projectPath: projectPath)
-
+        
+        let got = subject.targetTestAction(target: target,
+                                           pbxTarget: pbxTarget,
+                                           projectPath: projectPath)
+        
         XCTAssertEqual(got?.buildConfiguration, "Debug")
         XCTAssertEqual(got?.shouldUseLaunchSchemeArgsEnv, true)
         XCTAssertNil(got?.macroExpansion)
         XCTAssertEqual(got?.testables.count, 0)
     }
 
-    func test_testAction_when_testsTarget() {
+    func test_targetTestAction_when_testsTarget() {
         let target = Target.test(name: "AppTests", product: .unitTests)
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
 
-        let got = subject.testAction(target: target,
+        let got = subject.targetTestAction(target: target,
                                      pbxTarget: pbxTarget,
                                      projectPath: projectPath)
 
@@ -52,12 +115,12 @@ final class SchemeGeneratorTests: XCTestCase {
         XCTAssertEqual(buildableReference?.buildableIdentifier, "primary")
     }
 
-    func test_buildAction() {
+    func test_targetBuildAction() {
         let target = Target.test(name: "App", product: .app)
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
 
-        let got = subject.buildAction(target: target,
+        let got = subject.targetBuildAction(target: target,
                                       pbxTarget: pbxTarget,
                                       projectPath: projectPath)
 
@@ -75,11 +138,11 @@ final class SchemeGeneratorTests: XCTestCase {
         XCTAssertEqual(got?.buildImplicitDependencies, true)
     }
 
-    func test_launchAction_when_runnableTarget() {
+    func test_targetLaunchAction_when_runnableTarget() {
         let target = Target.test(name: "App", product: .app, environment: ["a": "b"])
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
-        let got = subject.launchAction(target: target,
+        let got = subject.targetLaunchAction(target: target,
                                        pbxTarget: pbxTarget,
                                        projectPath: projectPath)
 
@@ -94,13 +157,13 @@ final class SchemeGeneratorTests: XCTestCase {
         XCTAssertEqual(buildableReference?.buildableIdentifier, "primary")
     }
 
-    func test_launchAction_when_notRunnableTarget() {
+    func test_targetLaunchAction_when_notRunnableTarget() {
         let target = Target.test(name: "Library",
                                  platform: .iOS,
                                  product: .dynamicLibrary)
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
-        let got = subject.launchAction(target: target,
+        let got = subject.targetLaunchAction(target: target,
                                        pbxTarget: pbxTarget,
                                        projectPath: projectPath)
 
@@ -113,13 +176,13 @@ final class SchemeGeneratorTests: XCTestCase {
         XCTAssertEqual(got?.macroExpansion?.buildableIdentifier, "primary")
     }
 
-    func test_profileAction_when_runnableTarget() {
+    func test_targetProfileAction_when_runnableTarget() {
         let target = Target.test(name: "App",
                                  platform: .iOS,
                                  product: .app)
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
-        let got = subject.profileAction(target: target,
+        let got = subject.targetProfileAction(target: target,
                                         pbxTarget: pbxTarget,
                                         projectPath: projectPath)
 
@@ -145,13 +208,13 @@ final class SchemeGeneratorTests: XCTestCase {
         XCTAssertEqual(got?.enableTestabilityWhenProfilingTests, true)
     }
 
-    func test_profileAction_when_notRunnableTarget() {
+    func test_targetProfileAction_when_notRunnableTarget() {
         let target = Target.test(name: "Library",
                                  platform: .iOS,
                                  product: .dynamicLibrary)
         let pbxTarget = PBXNativeTarget(name: "App")
         let projectPath = AbsolutePath("/project.xcodeproj")
-        let got = subject.profileAction(target: target,
+        let got = subject.targetProfileAction(target: target,
                                         pbxTarget: pbxTarget,
                                         projectPath: projectPath)
 
@@ -177,14 +240,23 @@ final class SchemeGeneratorTests: XCTestCase {
         XCTAssertEqual(got?.macroExpansion?.buildableIdentifier, "primary")
     }
 
-    func test_analyzeAction() {
-        let got = subject.analyzeAction()
+    func test_targetAnalyzeAction() {
+        let got = subject.targetAnalyzeAction()
         XCTAssertEqual(got.buildConfiguration, "Debug")
     }
 
-    func test_archiveAction() {
-        let got = subject.archiveAction()
+    func test_targetArchiveAction() {
+        let got = subject.targetArchiveAction()
         XCTAssertEqual(got.buildConfiguration, "Release")
         XCTAssertEqual(got.revealArchiveInOrganizer, true)
     }
+    
+    // MARK: - Private
+    
+    private func generatedProject(targets: [Target]) -> GeneratedProject {
+        var pbxTargets: [String: PBXNativeTarget] = [:]
+        targets.forEach { pbxTargets[$0.name] = PBXNativeTarget(name: $0.name) }
+        let projectPath = AbsolutePath("/project.xcodeproj")
+        return GeneratedProject(path: projectPath, targets: pbxTargets)
+    }   
 }